+++ test/Braintree.TestUtil/MerchantAccountIDs.cs
@@ -13,5 +13,6 @@ namespace Braintree.TestUtil
         public static string ADYEN_MERCHANT_ACCOUNT_ID = "adyen_ma";
         public static string BRAZIL_MERCHANT_ACCOUNT_ID = "hiper_brl";
         public static string CARD_PROCESSOR_BRAZIL_MERCHANT_ACCOUNT_ID = "card_processor_brl";
+        public static string FAKE_FIRST_DATA_MERCHANT_ACCOUNT_ID = "fake_first_data_merchant_account";
     }
 }